Use correct domain
Prepares the production deployment for the correct domain.

In preparation for that to work, we need to set the DNS entry for `directory.btcpayserver.org` to a `CNAME` record pointing to `btcpayserver.github.io`.

See the [GH Pages docs on "Configuring a subdomain"](https://help.github.com/en/github/working-with-github-pages/managing-a-custom-domain-for-your-github-pages-site#configuring-a-subdomain).